feat: TechQuestionClient 클래스 새로 추가 및 기존 파일 삭제
